+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version 0.3.0
+
+* Add specific implementations for `exp2` and `log2` functions.
+
 ## Version 0.2.1
 
 * Added `exp_m1`, `ln_1p` functions.